TASK#16: Refactoring Deletes Unknown Action module Deletes unused types created for useDispatch and useSelector Mocks muted property of video element
