+++ project/src/setupTests.ts
@@ -3,17 +3,26 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
+import { Actions } from './store/store';
 
 const play = jest.fn().mockReturnValue(Promise.resolve());
 const load = jest.fn();
 const pause = jest.fn();
 
+const unknownAction = (): Actions => ({ type: 'UNKNOWN_ACTION'} as unknown as Actions);
+
 window.HTMLMediaElement.prototype.load = load;
 window.HTMLMediaElement.prototype.play = play;
 window.HTMLMediaElement.prototype.pause = pause;
 
+Object.defineProperty(window.HTMLMediaElement.prototype, 'muted', {
+  get: () => false,
+  set: jest.fn(),
+});
+
 export {
   play,
   load,
-  pause
+  pause,
+  unknownAction
 };
