--- project/src/store/unknown-action.ts
@@ -1,12 +1,0 @@
-enum UnknownActionType {
-  Unknown = 'data/unknown',
-}
-
-const unknownAction = () => ({
-  type: UnknownActionType.Unknown,
-} as const);
-
-export {
-  UnknownActionType,
-  unknownAction
-};
